fix(code smells)

+++ src/main/infra/adapters/operators/LogsRollupOperator.java
@@ -3,12 +3,10 @@ package main.infra.adapters.operators;
 import java.util.Calendar;
 import java.util.Properties;
 import java.util.concurrent.Callable;
-
 import org.json.JSONObject;
 import org.springframework.context.ApplicationContext;
-import org.springframework.web.context.ContextLoaderListener;
+import org.springframework.web.context.ContextLoader;
 import org.springframework.web.context.support.WebApplicationContextUtils;
-
 import main.domain.annotations.Operator;
 import main.infra.adapters.input.graphql.types.OperatorStage;
 import main.infra.adapters.output.repositories.SchedulerRepository;
@@ -20,16 +18,21 @@ public class LogsRollupOperator extends OperatorStage implements Callable<Void> 
 	@Override
 	public Void call() throws Exception {		
 		log.debug(this.getClass()+" init "+this.name);
-		ApplicationContext springContext = WebApplicationContextUtils.getWebApplicationContext(ContextLoaderListener.getCurrentWebApplicationContext().getServletContext());
-		SchedulerRepository repo = this.getSchedulerRepository(springContext);
 		var prop = new Properties();
-		prop.load(springContext.getClassLoader().getResourceAsStream("application.properties"));
-		
-		Calendar rollup = Calendar.getInstance();
-		rollup.setTimeInMillis(rollup.getTimeInMillis());
-		rollup.add(Calendar.HOUR, Integer.parseInt(prop.getProperty("param.logs.rollup.hours")));
-		repo.deleteLogsBy(rollup.getTime());
-		log.debug(this.getClass()+" end "+this.name);
+		var ctx = ContextLoader.getCurrentWebApplicationContext();
+		if(ctx != null) {
+			var srv = ctx.getServletContext();
+			if(srv != null) {
+				ApplicationContext springContext = WebApplicationContextUtils.getWebApplicationContext(srv);
+				SchedulerRepository repo = this.getSchedulerRepository(springContext);
+				prop.load(springContext.getClassLoader().getResourceAsStream("application.properties"));
+				Calendar rollup = Calendar.getInstance();
+				rollup.setTimeInMillis(rollup.getTimeInMillis());
+				rollup.add(Calendar.HOUR, Integer.parseInt(prop.getProperty("param.logs.rollup.hours")));
+				repo.deleteLogsBy(rollup.getTime());
+				log.debug(this.getClass()+" end "+this.name);	
+			} else return null;
+		}
 		return null;
 	}
 	@Override
